+++ src/Universum/Functor/Fmap.hs
@@ -1,32 +1,20 @@
 {-# LANGUAGE Safe #-}
 
+{-
+Copyright: (c) 2016 Stephen Diehl
+           (c) 20016-2018 Serokell
+           (c) 2018 Kowainik
+License: MIT
+-}
+
 -- | This module contains useful functions to work with 'Functor' type class.
 
 module Universum.Functor.Fmap
-       ( map
-       , (<<$>>)
+       ( (<<$>>)
        ) where
 
 import Universum.Function ((.))
 import Universum.Functor.Reexport (Functor (..))
-
--- $setup
--- >>> import Universum.Base (negate)
--- >>> import Universum.Bool (Bool (..), not)
--- >>> import Universum.Lifted (getLine)
--- >>> import Universum.Monad (Maybe (..))
--- >>> import Universum.String (toString)
-
-{- | 'Prelude.map' generalized to 'Functor'.
-
->>> map not (Just True)
-Just False
->>> map not [True,False,True,True]
-[False,True,False,False]
-
--}
-map :: Functor f => (a -> b) -> f a -> f b
-map = fmap
 
 -- $setup
 -- >>> import Universum.Base (negate)