+++ src/Universum/String/Reexport.hs
@@ -6,7 +6,6 @@ module Universum.String.Reexport
 
          -- * Text
        , module Data.Text
-       , module Data.Text.Lazy
        , module Data.Text.Encoding
        , module Data.Text.Encoding.Error
        , module Text.Read
@@ -21,5 +20,4 @@ import Data.Text (Text, lines, unlines, unwords, words)
 import Data.Text.Encoding (decodeUtf8', decodeUtf8With)
 import Data.Text.Encoding.Error (OnDecodeError, OnError, UnicodeException, lenientDecode,
                                  strictDecode)
-import Data.Text.Lazy (fromStrict, toStrict)
 import Text.Read (Read, readMaybe, reads)